New .csv functions, some internal BeachHandler changes
When parsing a .csv, the game can now increment a room "handler's" "localState" variable upon deletion, for e.g. scripting a cutscene in the Beach. It can also initiate a room change.

And of course, the localState variable changed how BeachHandler worked a tad - it's not a bit more sensible.